refactor: move types out of utils
