+++ src/utils/addSchemaToMetaData.ts
@@ -3,14 +3,16 @@ import path from 'path';
 import namify from 'namify';
 import filenamify from 'filenamify';
 import {
+  replaceInlinedRefsWithStringPlaceholder,
+  patchJsonSchema,
+  refToPath,
+} from '.';
+import type {
   SchemaMetaDataMap,
   SchemaMetaData,
   JSONSchema,
-  replaceInlinedRefsWithStringPlaceholder,
-  patchJsonSchema,
   SchemaPatcher,
-  refToPath,
-} from '.';
+} from '../types';
 
 /*
  * Just an utility function to add entries to SchemaMetaDataMap Map keyed by ref
