+++ src/utils/index.ts
@@ -13,12 +13,5 @@ export {
 } from './refReplacementUtils';
 export { replaceInlinedRefsWithStringPlaceholder } from './replaceInlinedRefsWithStringPlaceholder';
 export { replacePlaceholdersWithImportedSchemas } from './jsonSchemaToTsConst/replacePlaceholdersWithImportedSchemas';
-export type {
-  JSONSchema,
-  OpenApiSchema,
-  SchemaPatcher,
-  SchemaMetaData,
-  SchemaMetaDataMap,
-} from './types';
 export { addSchemaToMetaData } from './addSchemaToMetaData';
 export { makeRelativePath } from './makeRelativePath';
